llm_with_db/pyproject.toml
[tool.poetry]
name = "llm-with-db"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
diagrams = "^0.23.4"
azure-cosmos = "^4.6.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
